+++ src/components/layout/header/Header.tsx
@@ -46,7 +46,7 @@ const Header = () => {
         backgroundColor: isActive ? 'rgba(255, 255, 255, 0.9)' : 'transparent',
         clipPath: isActive ?
             'polygon(-10% -10%, 110% -10%, 110% 1000%, -10% 1000%)' :
-            'polygon(40% 0%, 60% 0%, 60% 100%, 40% 100%)',
+            'polygon(35% -15%, 65% -15%, 65% 100%, 35% 100%)',
         borderRadius: 100,
         mx: 'auto',
         left: 0,
